myapp/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
alembic==1.12.0
python-dotenv==1.0.0
redis==5.0.0
fastapi-cache2==0.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6